Added verify_files script
Created verify_files.py. This script attempts to process text files in a similar way to how the app processes the text. The script looks for errors and exceptions raised during this process and displays them to the user. This script is to be run prior to committing stories to the app’s data.